+++ app/resources/views/pages/dominion/government.blade.php
@@ -527,7 +527,7 @@
                                 <li>Enables war operations between members.</li>
                                 <li>75% of casualties suffered due to failed operations between members are automatically re-trained.</li>
                                 <li>Hourly infamy decay is reduced by 25%.</li>
-                                <li>Info op strength costs are halved (XP gain reduced by 25%).</li>
+                                <li>Info op strength costs are halved.</li>
                             </ul>
                             @if ($isLeavingBlackGuard)
                                 <form action="{{ route('dominion.government.black-guard.cancel') }}" method="post" role="form">