+++ app/resources/views/pages/dominion/op-center/show.blade.php
@@ -144,6 +144,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.magic') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('clear_sight'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'clear_sight']) }}" data-toggle="tooltip" title="Request a Clear Sight">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="spell" value="clear_sight">
                                 <button type="submit" class="btn btn-sm btn-primary">Clear Sight ({{ number_format($spellCalculator->getManaCost($selectedDominion, $infoSpells->get('clear_sight'))) }} mana)</button>
@@ -336,6 +345,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.magic') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('revelation'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'revelation']) }}" data-toggle="tooltip" title="Request a Revelation">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="spell" value="revelation">
                                 <button type="submit" class="btn btn-sm btn-primary">Revelation ({{ number_format($spellCalculator->getManaCost($selectedDominion, $infoSpells->get('revelation'))) }} mana)</button>
@@ -386,6 +404,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.espionage') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('castle_spy'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'castle_spy']) }}" data-toggle="tooltip" title="Request a Castle Spy">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="operation" value="castle_spy">
                                 <button type="submit" class="btn btn-sm btn-primary">Castle Spy</button>
@@ -439,6 +466,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.espionage') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('barracks_spy'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'barracks_spy']) }}" data-toggle="tooltip" title="Request a Barracks Spy">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="operation" value="barracks_spy">
                                 <button type="submit" class="btn btn-sm btn-primary">Barracks Spy</button>
@@ -510,6 +546,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.espionage') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('survey_dominion'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'survey_dominion']) }}" data-toggle="tooltip" title="Request a Survey Dominion">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="operation" value="survey_dominion">
                                 <button type="submit" class="btn btn-sm btn-primary">Survey Dominion</button>
@@ -579,6 +624,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.espionage') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('land_spy'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'land_spy']) }}" data-toggle="tooltip" title="Request a Land Spy">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="operation" value="land_spy">
                                 <button type="submit" class="btn btn-sm btn-primary">Land Spy</button>
@@ -649,6 +703,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.magic') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('vision'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'vision']) }}" data-toggle="tooltip" title="Request a Vision">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="spell" value="vision">
                                 <button type="submit" class="btn btn-sm btn-primary">Vision ({{ number_format($spellCalculator->getManaCost($selectedDominion, $infoSpells->get('vision'))) }} mana)</button>
@@ -748,6 +811,15 @@
                         <div class="pull-right">
                             <form action="{{ route('dominion.magic') }}" method="post" role="form">
                                 @csrf
+                                @if ($bounties->has('disclosure'))
+                                    <a data-toggle="tooltip" title="Op Requested">
+                                        <i class="fa fa-star text-yellow" style="margin-right: 10px;"></i>
+                                    </a>
+                                @else
+                                    <a href="{{ route('dominion.bounty-board.create', [$dominion->id, 'disclosure']) }}" data-toggle="tooltip" title="Request a Disclosure">
+                                        <i class="fa fa-star-o" style="margin-right: 10px;"></i>
+                                    </a>
+                                @endif
                                 <input type="hidden" name="target_dominion" value="{{ $dominion->id }}">
                                 <input type="hidden" name="spell" value="disclosure">
                                 <button type="submit" class="btn btn-sm btn-primary">Disclosure ({{ number_format($spellCalculator->getManaCost($selectedDominion, $infoSpells->get('disclosure'))) }} mana)</button>