+++ app/resources/views/partials/main-sidebar.blade.php
@@ -112,6 +112,16 @@
                 </li>
                 <li class="{{ Route::is('dominion.espionage') ? 'active' : null }}"><a href="{{ route('dominion.espionage') }}"><i class="fa fa-user-secret fa-fw"></i> <span>Espionage</span></a></li>
                 <li class="{{ Route::is('dominion.op-center*') ? 'active' : null }}"><a href="{{ route('dominion.op-center') }}"><i class="fa fa-globe fa-fw"></i> <span>Op Center</span></a></li>
+                <li class="{{ Route::is('dominion.bounty-board') ? 'active' : null }}">
+                    <a href="{{ route('dominion.bounty-board') }}">
+                        <i class="ra ra-hanging-sign ra-fw"></i> <span>Bounty Board</span>
+                        @if ($activeBounties > 0)
+                            <span class="pull-right-container">
+                                <small class="label label-primary pull-right">{{ $activeBounties }}</small>
+                            </span>
+                        @endif
+                    </a>
+                </li>
                 <li class="{{ Route::is('dominion.calculations') ? 'active' : null }}"><a href="{{ route('dominion.calculations') }}"><i class="fa fa-calculator fa-fw"></i> <span>Calculators</span></a></li>
 
                 <li class="header">RELATIONS</li>