+++ src/Providers/ComposerServiceProvider.php
@@ -10,6 +10,7 @@ use OpenDominion\Calculators\Dominion\Actions\TechCalculator;
 use OpenDominion\Calculators\Dominion\LandCalculator;
 use OpenDominion\Calculators\NetworthCalculator;
 use OpenDominion\Helpers\NotificationHelper;
+use OpenDominion\Models\Bounty;
 use OpenDominion\Models\MessageBoard;
 use OpenDominion\Services\Dominion\ProtectionService;
 use OpenDominion\Services\Dominion\SelectorService;
@@ -94,7 +95,6 @@ class ComposerServiceProvider extends AbstractServiceProvider
                 ->get([
                     'cast_by_dominion_id'
                 ]);
-
             $activeSelfSpells = 0;
             $activeHostileSpells = 0;
             foreach($activeSpells as $activeSpell) {
@@ -105,7 +105,6 @@ class ComposerServiceProvider extends AbstractServiceProvider
                     $activeHostileSpells++;
                 }
             }
-
             $view->with('activeSelfSpells', $activeSelfSpells);
             $view->with('activeHostileSpells', $activeHostileSpells);
 
@@ -115,23 +114,27 @@ class ComposerServiceProvider extends AbstractServiceProvider
             $unlockableTechCount = floor($selectedDominion->resource_tech / $techCost);
             $view->with('unlockableTechCount', $unlockableTechCount);
 
+            // Show barren land count
             $landCalculator = app(LandCalculator::class);
-
             $barrenLand = $landCalculator->getTotalBarrenLand($selectedDominion);
             $view->with('barrenLand', $barrenLand);
+
+            $activeBounties = Bounty::active()
+                ->where('source_realm_id', $selectedDominion->realm_id)
+                ->where('source_dominion_id', '!=', $selectedDominion->id)
+                ->count();
+            $view->with('activeBounties', $activeBounties);
 
             $unseenWonders = DB::table('round_wonders')
                 ->where('round_id', $selectedDominion->round_id)
                 ->where('created_at', '>', $selectedDominion->wonders_last_seen ?? $selectedDominion->round->created_at)
                 ->count();
-
             $view->with('unseenWonders', $unseenWonders);
 
             $unseenGameEvents = DB::table('game_events')
                 ->where('round_id', $selectedDominion->round_id)
                 ->where('created_at', '>', $selectedDominion->town_crier_last_seen ?? $selectedDominion->round->created_at)
                 ->count();
-
             $view->with('unseenGameEvents', $unseenGameEvents);
         });