+++ src/helpers.php
@@ -83,7 +83,7 @@ if (!function_exists('dominion_attr_display')) {
             'resource_lumber' => 'lumber',
             'resource_mana' => 'mana',
             'resource_ore' => 'ore',
-            'resource_tech' => 'tech',
+            'resource_tech' => 'research points',
             'land_water' => 'water',
         ];